Move Indent trait to generator::cpp::utils
