+++ crates/cxx-qt-gen/src/generator/cpp/mod.rs
@@ -73,6 +73,8 @@ impl GeneratedCppBlocks {
     }
 }
 
+mod utils;
+
 #[cfg(test)]
 mod tests {
     use super::*;
